back/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
scikit-learn==0.24.2
opencv-python==4.5.3.56
pillow==8.3.2
python-dotenv==0.19.0 